requirements.txt
asgiref>=3.8.1,<4
certifi==2023.7.22
Django==5.1
django-elasticsearch-dsl==8.0
elasticsearch==8.6.2
elasticsearch-dsl==8.17.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.16
python-decouple==3.8
django-crispy-forms==2.3
crispy-bootstrap5==2025.4
PyPDF2==3.0.1
python-docx==1.1.2
python-magic==0.4.27